remove /api and /v1 folders and move their files into /routes folder; configure winston logger in /src/config folder and fix import paths
