+++ src/app.ts
@@ -4,23 +4,10 @@ import express from 'express';
 import passport from 'passport';
 import cors from 'cors';
 import session from './config/session';
-import v1Router from './routes/api/v1/index';
+import apiRouter from './routes/index';
 import errorHandler from './middlewares/errorHandler';
-import winston from "winston";
 
 const app = express();
-
-export const logger = winston.createLogger({
-	level: "info",
-	format: winston.format.combine(
-		winston.format.timestamp(),
-		winston.format.json()
-	),
-	transports: [
-		new winston.transports.Console(),
-		new winston.transports.File({ filename: "errors.log", level: "error" })
-	]
-});
 
 app.use(express.json());
 app.enable('trust proxy');
@@ -31,7 +18,7 @@ app.use(cors({
 app.use(session);
 app.use(passport.initialize());
 app.use(passport.session());
-app.use('/api/v1', v1Router);
+app.use('/api', apiRouter);
 app.use(errorHandler);
 
 export default app;
