+++ src/config/logger.ts
@@ -1,0 +1,13 @@
+import winston from "winston";
+
+export const logger = winston.createLogger({
+	level: "info",
+	format: winston.format.combine(
+		winston.format.timestamp(),
+		winston.format.json()
+	),
+	transports: [
+		new winston.transports.Console(),
+		new winston.transports.File({ filename: "errors.log", level: "error" })
+	]
+});
