+++ src/routes/ai.ts
@@ -1,5 +1,5 @@
 import { RequestHandler, Router } from 'express';
-import aiController from '../../../controllers/aiController';
+import aiController from '../controllers/aiController';
 
 const {
 	enhanceText,
