+++ src/routes/auth.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import authController from '../../../controllers/authControllers';
+import authController from '../controllers/authControllers';
 
 const {
 	passportAuth,
