+++ src/routes/index.ts
@@ -3,16 +3,16 @@ import authRouter from './auth';
 import aiRouter from './ai';
 import userRouter from './me';
 import projectRouter from './projects/index';
-import isAuthenticated from '../../../middlewares/isAuthenticated';
+import isAuthenticated from '../middlewares/isAuthenticated';
 import checkRedisAndUseLimiter from '@/middlewares/aiLimiter';
 import timerRouter from './projects/timer';
 
 const router = Router();
 
-router.use('/auth', authRouter);
-router.use('/ai', isAuthenticated, checkRedisAndUseLimiter as RequestHandler, aiRouter);
-router.use('/me', isAuthenticated, userRouter);
-router.use('/projects', isAuthenticated, projectRouter);
-router.use('/timer', timerRouter);
+router.use('/v1/auth', authRouter);
+router.use('/v1/ai', isAuthenticated, checkRedisAndUseLimiter as RequestHandler, aiRouter);
+router.use('/v1/me', isAuthenticated, userRouter);
+router.use('/v1/projects', isAuthenticated, projectRouter);
+router.use('/v1/timer', timerRouter);
 
 export default router;
