+++ src/routes/me.ts
@@ -1,5 +1,5 @@
 import { Router, RequestHandler } from 'express';
-import userController from '../../../controllers/userController';
+import userController from '../controllers/userController';
 
 const {
     getMe,
